Trying to send a frame from C to C#

+++ src/utils/ppu_utils.c
@@ -11,7 +11,7 @@ void log_tile(tile *tile) {
 	}
 }
 
-void encode_as_tiles(byte *mem_addr, byte number_tiles, tile *tiles) {
+void encode_as_tiles(byte *mem_addr, uint number_tiles, tile *tiles) {
 	for (int i, tile_count = i = 0;
 	     tile_count < number_tiles; i += 16, ++tile_count) { //Each tile is defined by 16 bytes
 		for (int j = 0; j < TILE_ROW_SIZE; ++j) { //And each byte defines a row of pixels with a value from 0 to 3
